Move domain classes to data module to have domain unaware base module
