+++ client-data/src/main/java/com/jraska/github/client/users/UserDetail.kt
